Backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.4.0
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
filelock==3.13.1
idna==3.6
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.0